requirements.txt
Django==1.11.2
djangorestframework==3.6.3
djangorestframework-camel-case==0.2.0
drf-nested-routers==0.90.0
gevent==1.2.2
greenlet==0.4.12
pytz==2017.2
